Access window(s) element from pid
* This make sure we fully initialized the accessibility tree
