+++ macos/Onit/General/Extensions/ProcessIdentifier+Helper.swift
@@ -1,0 +1,18 @@
+//
+//  ProcessIdentifier+Helper.swift
+//  Onit
+//
+//  Created by Kévin Naudin on 13/05/2025.
+//
+
+import AppKit
+
+extension pid_t {
+    func getAppName() -> String? {
+        NSRunningApplication(processIdentifier: self)?.localizedName
+    }
+    
+    var bundleIdentifier: String? {
+        NSRunningApplication(processIdentifier: self)?.bundleIdentifier
+    }
+}
